+++ custom_components/meraki_ha/const_api.py
@@ -2,107 +2,5 @@
 
 from __future__ import annotations
 
-from typing import Final
-
-MERAKI_CONTENT_FILTERING_CATEGORIES: Final[list[dict[str, str]]] = [
-    {
-        "id": "meraki:contentFiltering/category/1",
-        "name": "Adult and Pornography",
-        "description": "Sites featuring sexual content, nudity, or pornography.",
-    },
-    {
-        "id": "meraki:contentFiltering/category/2",
-        "name": "Illegal",
-        "description": "Sites promoting illegal activities.",
-    },
-    {
-        "id": "meraki:contentFiltering/category/3",
-        "name": "Gambling",
-        "description": "Online gambling sites.",
-    },
-    {
-        "id": "meraki:contentFiltering/category/4",
-        "name": "Hate and Racism",
-        "description": "Sites promoting hatred or discrimination.",
-    },
-    {
-        "id": "meraki:contentFiltering/category/5",
-        "name": "Weapons",
-        "description": "Sites related to the sale or promotion of weapons.",
-    },
-    {
-        "id": "meraki:contentFiltering/category/6",
-        "name": "Violence",
-        "description": "Sites with graphic or gratuitous violence.",
-    },
-    {
-        "id": "meraki:contentFiltering/category/7",
-        "name": "Peer-to-peer",
-        "description": "Peer-to-peer file sharing sites and applications.",
-    },
-    {
-        "id": "meraki:contentFiltering/category/8",
-        "name": "Malware sites",
-        "description": "Sites known to host or distribute malware.",
-    },
-    {
-        "id": "meraki:contentFiltering/category/9",
-        "name": "Phishing and other frauds",
-        "description": "Sites engaged in phishing, scams, or other frauds.",
-    },
-    {
-        "id": "meraki:contentFiltering/category/10",
-        "name": "Key loggers and monitoring",
-        "description": "Sites for keyloggers, spyware, and monitoring tools.",
-    },
-    {
-        "id": "meraki:contentFiltering/category/11",
-        "name": "Botnets",
-        "description": "Sites associated with botnet command and control servers.",
-    },
-    {
-        "id": "meraki:contentFiltering/category/12",
-        "name": "Spam URLs",
-        "description": "URLs frequently found in unsolicited email (spam).",
-    },
-    {
-        "id": "meraki:contentFiltering/category/13",
-        "name": "Auctions",
-        "description": "Online auction sites.",
-    },
-    {
-        "id": "meraki:contentFiltering/category/14",
-        "name": "Games",
-        "description": "Online gaming sites.",
-    },
-    {
-        "id": "meraki:contentFiltering/category/15",
-        "name": "Social Networking",
-        "description": "Social networking sites and applications.",
-    },
-    {
-        "id": "meraki:contentFiltering/category/16",
-        "name": "Web-based email",
-        "description": "Web-based email services.",
-    },
-    {
-        "id": "meraki:contentFiltering/category/17",
-        "name": "Internet communications",
-        "description": "Chat, instant messaging, and other communication platforms.",
-    },
-    {
-        "id": "meraki:contentFiltering/category/18",
-        "name": "Shareware and freeware",
-        "description": "Sites for downloading shareware and freeware.",
-    },
-    {
-        "id": "meraki:contentFiltering/category/19",
-        "name": "Web advertisements",
-        "description": "Sites primarily serving advertisements.",
-    },
-    {
-        "id": "meraki:contentFiltering/category/20",
-        "name": "Nudity",
-        "description": "Sites with non-pornographic nudity.",
-    },
-]
+# Re-export for backward compatibility
+from .const_data import MERAKI_CONTENT_FILTERING_CATEGORIES  # noqa: F401
