+++ custom_components/meraki_ha/core/entities/__init__.py
@@ -8,10 +8,7 @@ from homeassistant.helpers.device_registry import DeviceInfo
 from homeassistant.helpers.entity import Entity
 from homeassistant.helpers.update_coordinator import CoordinatorEntity
 
-from ...const import (
-    DOMAIN,
-    MANUFACTURER,
-)
+from ...const import DOMAIN, MANUFACTURER
 from ...coordinator import MerakiDataUpdateCoordinator
 
 _LOGGER = logging.getLogger(__name__)
