+++ custom_components/meraki_ha/meraki_select/__init__.py
@@ -9,7 +9,8 @@ from homeassistant.config_entries import ConfigEntry
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
-from ..const import CONF_ENABLE_VPN_MANAGEMENT, DOMAIN
+from ..const import DOMAIN
+from ..const_conf import CONF_ENABLE_VPN_MANAGEMENT
 from .meraki_content_filtering import MerakiContentFilteringSelect
 from .vpn import MerakiVpnSelect
