+++ custom_components/meraki_ha/number/__init__.py
@@ -6,7 +6,8 @@ from homeassistant.config_entries import ConfigEntry
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
-from ..const import DOMAIN, PLATFORM_NUMBER
+from ..const import DOMAIN
+from ..const_platform import PLATFORM_NUMBER
 from .setup_helpers import async_setup_numbers
 
 _LOGGER = logging.getLogger(__name__)
