+++ custom_components/meraki_ha/switch/__init__.py
@@ -6,7 +6,8 @@ from homeassistant.config_entries import ConfigEntry
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
-from ..const import DOMAIN, PLATFORM_SWITCH
+from ..const import DOMAIN
+from ..const_platform import PLATFORM_SWITCH
 from .setup_helpers import async_setup_switches
 
 _LOGGER = logging.getLogger(__name__)
