+++ tests/helpers/test_device_info_helpers.py
@@ -4,9 +4,7 @@ from unittest.mock import MagicMock
 
 import pytest
 
-from custom_components.meraki_ha.const import (
-    DOMAIN,
-)
+from custom_components.meraki_ha.const import DOMAIN
 from custom_components.meraki_ha.helpers.device_info_helpers import resolve_device_info
