+++ tests/meraki_select/test_vpn_select.py
@@ -8,7 +8,8 @@ from homeassistant.helpers import entity_registry as er
 from homeassistant.setup import async_setup_component
 from pytest_homeassistant_custom_component.common import MockConfigEntry
 
-from custom_components.meraki_ha.const import CONF_ENABLE_VPN_MANAGEMENT, DOMAIN
+from custom_components.meraki_ha.const import DOMAIN
+from custom_components.meraki_ha.const_conf import CONF_ENABLE_VPN_MANAGEMENT
 from custom_components.meraki_ha.types import MerakiVpn
 from tests.const import MOCK_NETWORK
