+++ tests/sensor/network/test_traffic_shaping_sensor.py
@@ -5,7 +5,7 @@ from unittest.mock import AsyncMock, MagicMock
 import pytest
 from homeassistant.const import EntityCategory
 
-from custom_components.meraki_ha.const import CONF_ENABLE_TRAFFIC_SHAPING
+from custom_components.meraki_ha.const_conf import CONF_ENABLE_TRAFFIC_SHAPING
 from custom_components.meraki_ha.discovery.service import DeviceDiscoveryService
 from custom_components.meraki_ha.types import MerakiNetwork
