+++ tests/switch/test_firewall_rule.py
@@ -4,9 +4,7 @@ from unittest.mock import MagicMock
 
 import pytest
 
-from custom_components.meraki_ha.const import (
-    CONF_ENABLE_FIREWALL_RULES,
-)
+from custom_components.meraki_ha.const_conf import CONF_ENABLE_FIREWALL_RULES
 from custom_components.meraki_ha.switch.firewall_rule import MerakiFirewallRuleSwitch
 from custom_components.meraki_ha.switch.setup_helpers import async_setup_switches
 from custom_components.meraki_ha.types import MerakiFirewallRule
