+++ tests/switch/test_vlan_dhcp.py
@@ -4,9 +4,7 @@ from unittest.mock import MagicMock
 
 import pytest
 
-from custom_components.meraki_ha.const import (
-    CONF_ENABLE_VLAN_MANAGEMENT,
-)
+from custom_components.meraki_ha.const_conf import CONF_ENABLE_VLAN_MANAGEMENT
 from custom_components.meraki_ha.switch.setup_helpers import async_setup_switches
 from custom_components.meraki_ha.switch.vlan_dhcp import MerakiVLANDHCPSwitch
 from custom_components.meraki_ha.types import MerakiNetwork
