+++ tests/test_config_flow.py
@@ -18,10 +18,10 @@ from homeassistant.core import HomeAssistant
 from homeassistant.data_entry_flow import FlowResultType
 from pytest_homeassistant_custom_component.common import MockConfigEntry
 
-from custom_components.meraki_ha.const import (
+from custom_components.meraki_ha.const import DOMAIN
+from custom_components.meraki_ha.const_conf import (
     CONF_MERAKI_API_KEY,
     CONF_MERAKI_ORG_ID,
-    DOMAIN,
 )
 from custom_components.meraki_ha.core.errors import (
     MerakiAuthenticationError,
