+++ tests/test_e2e_device_status.py
@@ -17,10 +17,10 @@ from playwright.async_api import Error as PlaywrightError
 from playwright.async_api import async_playwright, expect
 from pytest_homeassistant_custom_component.common import MockConfigEntry
 
-from custom_components.meraki_ha.const import (
+from custom_components.meraki_ha.const import DOMAIN
+from custom_components.meraki_ha.const_conf import (
     CONF_MERAKI_API_KEY,
     CONF_MERAKI_ORG_ID,
-    DOMAIN,
 )
 from custom_components.meraki_ha.types import MerakiDevice
